+ Enable promoting warnings to errors in CMake to trigger pipeline
  faults for these warnings going forward
+ Do not build OE module with no-asm (RdSeed support is essential
for OE)

Related work items: #38706387

+++ lib/sc_lib.h
@@ -1759,7 +1759,7 @@ VOID
 SYMCRYPT_CALL
 SymCryptKeccakAppendEncodeTimes8(
     _Inout_ SYMCRYPT_KECCAK_STATE *pState,
-            SIZE_T  uValue,
+            UINT64  uValue,
             BOOLEAN bLeftEncode);
 // Appends the left-encoding of uValue * 8 to the state
